Split kraken and kraken-web.
