+++ kraken-web/src/Kraken/Web/Config.hs
