+++ kraken-web/test/Kraken/Web/ConfigSpec.hs
