+++ kraken/Setup.hs
