+++ kraken/examples/Example.hs
