+++ kraken/kraken.cabal
@@ -1,0 +1,144 @@
+name:             kraken
+version:          0.2.0.0
+author:           Sönke Hahn
+maintainer:       soenke.hahn@zalora.com
+copyright:        Zalora South East Asia Pte Ltd
+license:          BSD3
+license-file:     LICENSE
+category:         Database
+build-type:       Simple
+cabal-version:    >= 1.10
+
+extra-source-files:
+  default.nix
+
+data-files:
+  static/*.html
+
+flag network-uri
+   description: Get Network.URI from the network-uri package
+   default: True
+
+library
+  hs-source-dirs:
+      src
+  build-depends:
+      base == 4.*
+    , aeson
+    , bytestring
+    , configurator
+    , containers
+    , directory
+    , either == 4.3.*
+    , enclosed-exceptions
+    , filepath
+    , graph-wrapper
+    , http-client
+    , http-types
+    , interpolate
+    , mtl
+    , optparse-applicative >= 0.11
+    , process
+    , safe
+    , servant
+    , servant-server
+    , servant-docs
+    , servant-client
+    , string-conversions
+    , system-filepath
+    , text
+    , transformers
+    , wai
+    , wai-app-static
+    , warp
+    , yaml
+  if flag(network-uri)
+    build-depends: network-uri >= 2.6, network >= 2.6
+  else
+    build-depends: network-uri < 2.6, network < 2.6
+  exposed-modules:
+      Kraken
+      Kraken.ActionM
+      Kraken.Config
+      Kraken.Daemon
+      Kraken.Dot
+      Kraken.Graph
+      Kraken.Run
+      Kraken.Store
+      Kraken.Util
+      Kraken.TargetGraph
+      --Kraken.Web
+      --Kraken.Web.Config
+      --Kraken.Web.Utils
+      --Kraken.Web.TargetGraph
+      Network.Wai.Handler.Warp.Run
+      Paths_kraken
+  default-language:
+      Haskell2010
+  ghc-options:
+      -Wall -fno-warn-name-shadowing
+      -threaded
+
+test-suite krakenspec
+  type:
+      exitcode-stdio-1.0
+  ghc-options:
+      -Wall -fno-warn-name-shadowing
+      -threaded
+      -O0
+  default-language:
+      Haskell2010
+  hs-source-dirs:
+      test, examples
+  main-is:
+      Spec.hs
+  other-modules:
+    ExampleSpec
+    Spec
+    Test.Hspec.Checkers
+    Kraken.ActionMSpec
+    Kraken.StoreSpec
+    --Kraken.WebSpec
+    Kraken.GraphSpec
+    --Kraken.Web.ConfigSpec
+    Kraken.DaemonSpec
+    Kraken.DotSpec
+    Kraken.RunSpec
+    Kraken.ConfigSpec
+    KrakenSpec
+  build-depends:
+      base == 4.*
+    , aeson
+    , checkers
+    , configurator
+    , containers
+    , deepseq
+    , directory
+    , either
+    , graph-wrapper
+    , hspec
+    , hspec-wai >= 0.5
+    , http-client
+    , http-types
+    , HUnit
+    , interpolate
+    , kraken
+    , process
+    , QuickCheck >= 2.7
+    , quickcheck-instances
+    , safe
+    , servant
+    , servant-server
+    , servant-docs
+    , servant-client
+    , silently
+    , string-conversions
+    , temporary
+    , transformers
+    , wai
+    , wai-extra
+    , warp
+  if flag(network-uri)
+    build-depends: network-uri >= 2.6, network >= 2.6
+  else
+    build-depends: network-uri < 2.6, network < 2.6
