+++ kraken/src/Kraken/Graph.hs
