+++ kraken/src/Kraken/Util.hs
