+++ kraken/test/Kraken/ConfigSpec.hs
