+++ kraken/test/Kraken/DaemonSpec.hs
