+++ kraken/test/Kraken/DotSpec.hs
