+++ kraken/test/Kraken/GraphSpec.hs
