+++ kraken/test/Kraken/StoreSpec.hs
