+++ kraken/test/Test/Hspec/Checkers.hs
